refactor: migrate all helpers into lib folder
