+++ lib/commands.js
@@ -1,4 +1,3 @@
-const log = require('debug')('mongo:rs')
 const chalk = require('chalk')
 const rs = require('./rs')
