+++ lib/conf.js
