+++ lib/mongo.js
@@ -10,7 +10,7 @@ module.exports = async url => {
     await client.connect()
 
     return client
-  } catch(e) {
+  } catch (e) {
     console.error(e.stack)
     process.exit(1)
   }
