+++ lib/rs/index.js
@@ -3,5 +3,5 @@ const initiate = require('./initiate')
 
 module.exports = {
   status,
-  initiate,
+  initiate
 }
