+++ lib/rs/initiate.js
@@ -32,5 +32,4 @@ module.exports = async (mongo, replicaSet) => {
 
     throw e
   }
-
 }
