+++ lib/stdin.js
@@ -5,7 +5,7 @@ const chalk = require('chalk')
 readline.emitKeypressEvents(process.stdin)
 process.stdin.setRawMode(true)
 
-const keyMap = new Map();
+const keyMap = new Map()
 
 keyMap.set('s', 'status')
 keyMap.set('q', 'quit')
